+++ packages/header/tests/react/index.js
@@ -1,100 +1,175 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import AUheader from './header.js';
+import AUheader, { AUheaderBrand } from './header.js';
 
 
 ReactDOM.render(
 	<div>
 		<h2>header default</h2>
 
-		<AUheader title="Page title" />
+		<AUheader className="au-grid">
+			<div className="container">
+				<div className="row">
+					<div className="col-md-9">
+						<AUheaderBrand
+							link="#"
+							brandImage="http://placehold.it/256x80"
+							brandImageAlt="Digital Transformation Agency"
+						/>
+					</div>
+				</div>
+			</div>
+		</AUheader>
+
+		<h2>header <code>alt</code> color</h2>
+
+		<AUheader alt className="au-grid">
+			<div className="container">
+				<div className="row">
+					<div className="col-md-9">
+						<AUheaderBrand
+							link="#"
+							brandImage="http://placehold.it/256x80"
+							brandImageAlt="Digital Transformation Agency"
+						/>
+					</div>
+				</div>
+			</div>
+		</AUheader>
+
+		<h2>header <code>dark</code> color</h2>
+
+		<AUheader dark className="au-grid">
+			<div className="container">
+				<div className="row">
+					<div className="col-md-9">
+						<AUheaderBrand
+							link="#"
+							brandImage="http://placehold.it/256x80"
+							brandImageAlt="Digital Transformation Agency"
+						/>
+					</div>
+				</div>
+			</div>
+		</AUheader>
+
+		<h2>header <code>dark</code> <code>alt</code> color</h2>
+
+		<AUheader dark alt className="au-grid">
+			<div className="container">
+				<div className="row">
+					<div className="col-md-9">
+						<AUheaderBrand
+							link="#"
+							brandImage="http://placehold.it/256x80"
+							brandImageAlt="Digital Transformation Agency"
+						/>
+					</div>
+				</div>
+			</div>
+		</AUheader>
+
+		<hr />
+
+		<h2>header with title and subline</h2>
+
+		<AUheader dark alt className="au-grid">
+			<div className="container">
+				<div className="row">
+					<div className="col-md-9">
+						<AUheaderBrand
+							link="#"
+							title="Page title"
+							subline="Service sub-title"
+						/>
+					</div>
+				</div>
+			</div>
+		</AUheader>
+
+		<hr />
+
+		<h2>header with title and subline</h2>
+
+		<AUheader dark alt className="au-grid">
+			<div className="container">
+				<div className="row">
+					<div className="col-md-9">
+						<AUheaderBrand
+							link="#"
+							brandImage="http://placehold.it/256x80"
+							brandImageAlt="Digital Transformation Agency"
+							title="Page title"
+							subline="Service sub-title"
+						/>
+					</div>
+				</div>
+			</div>
+		</AUheader>
 
 
 		<hr />
 		<h2>header with additional class</h2>
 
-		<AUheader dark title="Page title" subline="Service sub-title" className="testing" />
+		<AUheader dark className="testing au-grid">
+			<div className="container">
+				<div className="row">
+					<div className="col-md-9">
+						<AUheaderBrand
+							title="Page title"
+							subline="Service sub-title"
+							link="#"
+							brandImage="http://placehold.it/256x80"
+							brandImageAlt="Digital Transformation Agency"
+						/>
+					</div>
+				</div>
+			</div>
+		</AUheader>
 
 
 		<hr />
 		<h2>header with children</h2>
 
-		<AUheader title="Section title">
-			<button>Menu button?</button>
+		<AUheader dark className="au-grid">
+			<div className="container">
+				<div className="row">
+					<div className="col-md-9">
+						<AUheaderBrand link="#" title="Page title" subline="Service sub-title" />
+					</div>
+				</div>
+			</div>
+			<p>Some navigation maybe</p>
 		</AUheader>
 
 
 		<hr />
 		<h2>header with different heading level</h2>
 
-		<AUheader title="Section title" level="2" />
+		<AUheader dark className="au-grid">
+			<div className="container">
+				<div className="row">
+					<div className="col-md-9">
+						<AUheaderBrand level="5" link="#" title="Page title" subline="Service sub-title" />
+					</div>
+				</div>
+			</div>
+		</AUheader>
 
 
 		<hr />
-		<h2>header with subline</h2>
+		<h2>header <code>--hero</code></h2>
 
-		<AUheader title="Page title" subline="Service sub-title that could be a little longer" />
-
-
-		<hr />
-		<h2>header <code>--hero</code> with subline</h2>
-
-		<AUheader title="Page title" hero subline="Service sub-title that could be a little longer" />
-
-
-		<hr />
-		<h2>header <code>--alt</code></h2>
-
-		<AUheader title="Page title" alt />
-
-
-		<hr />
-		<h2>header <code>--alt</code> with subline</h2>
-
-		<AUheader title="Page title" alt subline="Service sub-title that could be a little longer" />
-
-
-		<hr />
-		<h2>header  <code>--hero</code> <code>--alt</code> with subline</h2>
-
-		<AUheader title="Page title" alt hero subline="Service sub-title that could be a little longer" />
-
-
-		<hr />
-		<h2>header <code>--dark</code></h2>
-
-		<AUheader title="Page title" dark />
-
-
-		<hr />
-		<h2>header <code>--dark</code> with subline</h2>
-
-		<AUheader title="Page title" dark subline="Service sub-title that could be a little longer" />
-
-
-		<hr />
-		<h2>header <code>--hero</code> <code>--dark</code> with subline</h2>
-
-		<AUheader title="Page title" dark hero subline="Service sub-title that could be a little longer" />
-
-
-		<hr />
-		<h2>header <code>--alt</code> <code>--dark</code></h2>
-
-		<AUheader title="Page title" alt dark />
-
-
-		<hr />
-		<h2>header <code>--alt</code> <code>--dark</code> with subline</h2>
-
-		<AUheader title="Page title" alt dark subline="Service sub-title that could be a little longer" />
-
-
-		<hr />
-		<h2>header <code>--hero</code> <code>--alt</code> <code>--dark</code> with subline</h2>
-
-		<AUheader title="Page title" alt dark hero subline="Service sub-title that could be a little longer" />
+		<AUheader hero dark className="au-grid">
+			<div className="container">
+				<div className="row">
+					<div className="col-md-9">
+						<AUheaderBrand level="5" link="#" title="Page title" subline="Service sub-title" />
+					</div>
+				</div>
+			</div>
+		</AUheader>
 
 	</div>,