+++ packages/header/CHANGELOG.md
@@ -15,6 +15,7 @@
 
 ## Versions
 
+* [v4.0.0 - Added placeholder for Australian Government logo and optional grid layout ](#v400)
 * [v3.0.2 - Change homepage link ](#v302)
 * [v3.0.1 - Fix dependencies ](#v301)
 * [v3.0.0 - Change to focus colour and border/muted color mix](#v300)
@@ -29,6 +30,11 @@
 
 
 ## Release History
+
+### v4.0.0
+
+- Added placeholder for Australian Government logo and optional grid layout
+
 
 ### v3.0.2